[#200] :truck: rename: api instance 경로 변경
