+++ client/src/components/diary-revision/DiaryRevisionForm.tsx
@@ -6,7 +6,8 @@ import DiaryRevisionImgUpload from "./DiaryRevisionImgUpload";
 import { diaryRevisionImgAtom } from "src/recoil/diary-revision/atom";
 import { Img } from "src/recoil/new-diary/atom";
 import { API_URL } from "src/constants/API_URL";
-import { api, formApi } from "src/utils/axiosApi";
+import { api } from "src/utils/axiosApi/api";
+import { formApi } from "src/utils/axiosApi/formApi";
 import { useNavigate, useParams } from "react-router-dom";
 import { DiaryForm, TitleInput, inputStyle, ContentInput } from "../diary/DiaryFormElem";
 import { DiaryInput } from "../diary/DiaryInput";
