+++ client/src/components/edit-info/EditNickname.tsx
@@ -3,7 +3,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
 
 import { API_URL } from "src/constants/API_URL";
-import { api } from "src/utils/axiosApi";
+import { api } from "src/utils/axiosApi/api";
 import { PRIVATE_ROUTE } from "src/router/ROUTE_INFO";
 import PurpleButton from "src/components/common/PurpleButton";
 import { Input, InputContainer } from "src/components/common/Input";
