+++ client/src/components/edit-info/EditPassword.tsx
@@ -3,7 +3,7 @@ import tw from "tailwind-styled-components";
 import { useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
 
-import { api } from "src/utils/axiosApi";
+import { api } from "src/utils/axiosApi/api";
 import { API_URL } from "src/constants/API_URL";
 import { PRIVATE_ROUTE } from "src/router/ROUTE_INFO";
 import { Input, InputContainer } from "src/components/common/Input";
