+++ client/src/components/home/EmailLoginContainer.tsx
@@ -6,7 +6,7 @@ import { accessTokenAtom } from "src/recoil/token";
 import { API_URL } from "src/constants/API_URL";
 import { Input, InputContainer } from "../common/Input";
 import PurpleButton from "../common/PurpleButton";
-import { api } from "src/utils/axiosApi";
+import { api } from "src/utils/axiosApi/api";
 
 interface loginInput {
   readonly email: string;
